first pass
lets store user and game data in rethinkdb
when a game ui component is activated send a message via ws to the node
server, server tries to handle the game event, applying it to the
current state. If all is ok, send the hash of the game's current state.
The client can then check that hash against the hash of their own
version of the game's current state. BOOM.